backend/requirements.txt
flask~=3.0.3
mysql-connector-python~=9.0.0
sqlalchemy~=2.0.35